Mod 0 Task 1

+++ 0x15-file_io/1-create_file.c
@@ -11,7 +11,7 @@
 int create_file(const char *filename, char *text_content)
 {
 	int fd, count = 0;
-	size_t written;
+	int written;
 
 	if (filename == NULL)
 	{
